# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fuel Nutrition API",
    "version": "1.0.0",
    "description": "Nutrition tracking API for the Fuel AI coach. Logs food, tracks weight, manages goals, and provides coaching insights."
  },
  "servers": [
    {
      "url": "https://fuel-2j8v.onrender.com/api",
      "description": "Production"
    }
  ],
...
  "paths": {
...
    "/weight": {
      "get": {
        "tags": [
          "Weight"
        ],
        "summary": "Get weight trend data",
        "description": "Weight tracking view with trend series, smoothed stats, and headline metrics for a selected window.",
        "operationId": "getWeight",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                30,
                60,
                90,
                180
              ],
              "default": 90
            },
            "description": "Lookback window in days."
          }
        ],
        "responses": {
          "200": {
            "description": "Weight data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightData"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Weight"
        ],
        "summary": "Log a weigh-in",
        "description": "Record or overwrite a scale weight for a given date. Returns updated WeightData.",
        "operationId": "logWeight",
        "parameters": [
          {
            "name": "range",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "enum": [
                30,
                60,
                90,
                180
              ],
              "default": 90
            },
            "description": "Lookback window for the response."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "weight"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "ISO date (YYYY-MM-DD)."
                  },
                  "weight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "maximum": 500,
                    "description": "Scale weight in kg."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Weigh-in recorded — returns updated weight data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeightData"
                }
              }
            }
          },
          "400": {
            "description": "Invalid date or weight out of range",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/goal": {
      "get": {
        "tags": [
          "Goal"
        ],
        "summary": "Get current goal",
        "description": "Returns the active user goal configuration (mode, target weight, start weight, start date).",
        "operationId": "getGoal",
        "responses": {
          "200": {
            "description": "Goal configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoal"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Goal"
        ],
        "summary": "Set or update goal",
        "description": "Overwrite the active goal. Missing optional fields fall back to the existing values.",
        "operationId": "setGoal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "mode",
                  "targetWeight"
                ],
                "properties": {
                  "mode": {
                    "$ref": "#/components/schemas/GoalMode"
                  },
                  "targetWeight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "description": "Target body weight in kg."
                  },
                  "startWeight": {
                    "type": "number",
                    "exclusiveMinimum": 0,
                    "description": "Starting body weight in kg (optional, falls back to existing)."
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date",
                    "description": "ISO start date (optional, falls back to existing)."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Goal updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGoal"
                }
              }
            }
          },
          "400": {
            "description": "Invalid mode or targetWeight",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/training-day": {
      "get": {
        "tags": ["Goal"],
        "summary": "Get training day status",
        "description": "Returns whether a specific date is a training day. Training days get +250 kcal, +20g protein, more carbs. Defaults to demo anchor date.",
        "operationId": "getTrainingDay",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": false,
            "schema": { "type": "string", "format": "date" },
            "description": "ISO date (YYYY-MM-DD). Defaults to demo anchor."
          }
        ],
        "responses": {
          "200": {
            "description": "Training day status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["isTrainingDay"],
                  "properties": {
                    "isTrainingDay": { "type": "boolean" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Goal"],
        "summary": "Set or toggle training day",
        "description": "Set a specific date as training day or rest day. Omit isTraining to toggle. Training day bumps calorie/protein/carb targets by ~250 kcal / 20g protein / 45g carbs.",
        "operationId": "setTrainingDay",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "ISO date. Defaults to demo anchor."
                  },
                  "isTraining": {
                    "type": "boolean",
                    "description": "true = training day, false = rest day. Omit to toggle."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated training day status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["isTrainingDay", "date"],
                  "properties": {
                    "isTrainingDay": { "type": "boolean" },
                    "date": { "type": "string", "format": "date" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/measurements": {
      "get": {
        "tags": ["Progress"],
        "summary": "Get body measurements",
        "description": "Returns all logged body measurements (waist, chest, arms, hips, thigh, BF%) oldest→newest plus the latest snapshot.",
        "operationId": "getMeasurements",
        "responses": {
          "200": {
            "description": "Measurements list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "measurements": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Measurement" }
                    },
                    "latest": {
                      "oneOf": [
                        { "$ref": "#/components/schemas/Measurement" },
                        { "type": "null" }
                      ]
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Progress"],
        "summary": "Log a body measurement",
        "description": "Log circumference measurements (cm). If waist + neck + height are provided, body fat % is auto-estimated via the US Navy formula. Overwrites any existing entry for the same date.",
        "operationId": "logMeasurement",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "date": { "type": "string", "format": "date", "description": "ISO date. Default: 2026-07-02." },
                  "waist": { "type": "number", "description": "Waist circumference in cm." },
                  "hips": { "type": "number", "description": "Hip circumference in cm." },
                  "chest": { "type": "number", "description": "Chest circumference in cm." },
                  "armLeft": { "type": "number", "description": "Left arm circumference in cm." },
                  "armRight": { "type": "number", "description": "Right arm circumference in cm." },
                  "thigh": { "type": "number", "description": "Thigh circumference in cm." },
                  "neck": { "type": "number", "description": "Neck circumference in cm — needed for Navy BF% auto-calculation." },
                  "height": { "type": "number", "description": "Height in cm — needed for Navy BF% auto-calculation." },
                  "sex": { "type": "string", "enum": ["male", "female"], "description": "Sex for Navy BF% formula. Default: male." },
                  "bodyFat": { "type": "number", "description": "Body fat % if known directly (overrides auto-calculation)." }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Measurement logged",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Measurement" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/user/register": {
      "post": {
        "tags": ["User"],
        "summary": "Register a new user",
        "description": "Creates a new user ID (UUID). Store this ID and send it as X-User-Id header on all subsequent requests.",
        "operationId": "registerUser",
        "parameters": [],
        "responses": {
          "201": {
            "description": "New user created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["userId", "createdAt"],
                  "properties": {
                    "userId": { "type": "string", "format": "uuid", "description": "Unique user identifier. Send as X-User-Id header on all requests." },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/user": {
      "get": {
        "tags": ["User"],
        "summary": "Get current user info",
        "description": "Returns user info for the authenticated user. Requires X-User-Id header.",
        "operationId": "getUser",
        "parameters": [
          { "$ref": "#/components/parameters/UserIdHeader" }
        ],
        "responses": {
          "200": {
            "description": "User info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["userId", "createdAt"],
                  "properties": {
                    "userId": { "type": "string", "format": "uuid" },
                    "createdAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Error" },
          "404": { "$ref": "#/components/responses/Error" },
          "500": { "$ref": "#/components/responses/Error" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "UserIdHeader": {
        "name": "X-User-Id",
        "in": "header",
        "required": false,
        "schema": { "type": "string", "format": "uuid" },
        "description": "User identifier from POST /api/user/register. If omitted, uses the default shared user."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message."
          }
        }
      },
...
      "GoalMode": {
        "type": "string",
        "enum": [
          "fat-loss",
          "maintenance",
          "lean-bulk",
          "recomp"
        ],
        "description": "Nutrition goal mode."
      },
...
      "WeightPoint": {
        "type": "object",
        "required": [
          "date",
          "scale",
          "trend"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "scale": {
            "type": "number",
            "nullable": true,
            "description": "Raw scale weight (null if no weigh-in that day)."
          },
          "trend": {
            "type": "number",
            "description": "Smoothed trend weight."
          }
        }
      },
...
      "WeightStats": {
        "type": "object",
        "required": [
          "currentTrend",
          "latestScale",
          "weeklyRate",
          "totalChange",
          "lowestTrend",
          "highestTrend",
          "entriesLogged",
          "windowDays"
        ],
        "properties": {
          "currentTrend": {
            "type": "number",
            "description": "Latest smoothed trend weight (kg)."
          },
          "latestScale": {
            "type": "number",
            "nullable": true,
            "description": "Most recent raw scale weight (kg)."
          },
          "weeklyRate": {
            "type": "number",
            "description": "Trend change over last 7 days (kg, signed)."
          },
          "totalChange": {
            "type": "number",
            "description": "Trend change across full window (kg, signed)."
          },
          "lowestTrend": {
            "type": "number",
            "description": "Lowest trend weight in window (kg)."
          },
          "highestTrend": {
            "type": "number",
            "description": "Highest trend weight in window (kg)."
          },
          "entriesLogged": {
            "type": "integer",
            "description": "Days with a raw weigh-in in window."
          },
          "windowDays": {
            "type": "integer",
            "description": "Total days in the window."
          }
        }
      },
      "WeightData": {
        "type": "object",
        "required": [
          "date",
          "range",
          "series",
          "stats",
          "unit"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "description": "Date of the newest data point."
          },
          "range": {
            "type": "integer",
            "enum": [
              30,
              60,
              90,
              180
            ],
            "description": "Lookback window."
          },
          "series": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeightPoint"
            }
          },
          "stats": {
            "$ref": "#/components/schemas/WeightStats"
          },
          "unit": {
            "type": "string",
            "enum": [
              "kg",
              "lb"
            ],
            "description": "Weight unit."
          }
        }
      },
...
      "Measurement": {
        "type": "object",
        "required": [
          "date"
        ],
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "waist": {
            "type": "number",
            "nullable": true,
            "description": "cm"
          },
          "hips": {
            "type": "number",
            "nullable": true,
            "description": "cm"
          },
          "chest": {
            "type": "number",
            "nullable": true,
            "description": "cm"
          },
          "armLeft": {
            "type": "number",
            "nullable": true,
            "description": "cm"
          },
          "armRight": {
            "type": "number",
            "nullable": true,
            "description": "cm"
          },
          "thigh": {
            "type": "number",
            "nullable": true,
            "description": "cm"
          },
          "bodyFat": {
            "type": "number",
            "nullable": true,
            "description": "Body fat percentage."
          }
        }
      },
...
      "UserGoal": {
        "type": "object",
        "required": [
          "mode",
          "targetWeight",
          "startWeight",
          "startDate"
        ],
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/GoalMode"
          },
          "targetWeight": {
            "type": "number",
            "description": "Goal body weight in kg."
          },
          "startWeight": {
            "type": "number",
            "description": "Starting body weight in kg."
          },
          "startDate": {
            "type": "string",
            "format": "date",
            "description": "Goal start date."
          }
        }
      },
...
    }
  }
}